gatekeeper: add matchLabels selector to parameters (#24)
The label selector the controller uses to select remote services is
configurable, therefore I purposely omitted any label matching from the
policy itself, assuming that using a `labelSelector` on the constaint
itself would suffice.

As it turns out, the constraint `labelSelector` will still select the
object when the label is removed (because the old object matches the
selector). So, in order to support removing the the label from a
violating object, the policy needs to check the labels itself.